+++ Basil/include/Scene/Component.h
@@ -75,4 +75,31 @@ namespace Basil
 		glm::vec3 rotation = { 0.0f, 0.0f, 0.0f };
 		glm::vec3 scale{ 1.0f, 1.0f, 1.0f };
 	};
+
+	// Physics
+	struct Rigidbody2DComponent
+	{
+		enum class BodyType { Static = 0, Dynamic, Kinematic };
+
+		Rigidbody2DComponent() = default;
+		Rigidbody2DComponent(const Rigidbody2DComponent& other) = default;
+
+		BodyType type = BodyType::Static;
+		bool fixedRotation = false;
+		void* runtimeBody = nullptr;
+	};
+
+	struct BoxCollider2DComponent
+	{
+		BoxCollider2DComponent() = default;
+		BoxCollider2DComponent(const BoxCollider2DComponent& other) = default;
+
+		glm::vec2 offset = { 0.0f, 0.0f };
+		glm::vec2 size = { 0.0f, 0.0f };
+		float density = 1.0f;
+		float friction = 0.5f;
+		float restitution = 0.0f;
+		float restitutionThreshold = 0.5f;
+		void* runtimeBody = nullptr;
+	};
 }