+++ Basil/src/Scene/Scene.cpp
@@ -4,16 +4,32 @@
 #include "Scene/Entity.h"
 #include "Scene/Scene.h"
 
+#include <box2d/b2_body.h>
+#include <box2d/b2_fixture.h>
+#include <box2d/b2_polygon_shape.h>
+#include <box2d/b2_world.h>
 #include <glm/glm.hpp>
 
 namespace Basil
 {
+	// Convert Basil Rigidbody type to Box2D type
+	static b2BodyType BasilRigidbody2DTypeToBox2D(Rigidbody2DComponent::BodyType bodyType)
+	{
+		switch (bodyType)
+		{
+			case Rigidbody2DComponent::BodyType::Static:	return b2_staticBody;
+			case Rigidbody2DComponent::BodyType::Dynamic:	return b2_dynamicBody;
+			case Rigidbody2DComponent::BodyType::Kinematic:	return b2_kinematicBody;
+		}
+	}
+
 	// Constructor
 	Scene::Scene()
 	{
-		// Initialis variables
+		// Initialise variables
 		viewportWidth = 0;
 		viewportHeight = 0;
+		box2DWorld = nullptr;
 	}
 
 	// Destructor
@@ -43,6 +59,51 @@ namespace Basil
 		registry.destroy(entity);
 	}
 
+	// On runtime start
+	void Scene::onRuntimeStart()
+	{
+		box2DWorld = new b2World({ 0.0f, -9.8f });
+		auto view = registry.view<Rigidbody2DComponent>();
+		for (auto e : view)
+		{
+			Entity entity = { e, this };
+			auto& transform = entity.getComponent<TransformComponent>();
+			auto& rb2d = entity.getComponent<Rigidbody2DComponent>();
+
+			b2BodyDef bodyDef;
+			bodyDef.type = BasilRigidbody2DTypeToBox2D(rb2d.type);
+			bodyDef.position.Set(transform.translation.x, transform.translation.y);
+			bodyDef.angle = transform.rotation.z;
+
+			b2Body* body = box2DWorld->CreateBody(&bodyDef);
+			body->SetFixedRotation(rb2d.fixedRotation);
+			rb2d.runtimeBody = body;
+
+			if (entity.hasComponent<BoxCollider2DComponent>())
+			{
+				auto& bc2d = entity.getComponent<BoxCollider2DComponent>();
+
+				b2PolygonShape polygonShape;
+				polygonShape.SetAsBox(bc2d.size.x * transform.scale.x, bc2d.size.y * transform.scale.y);
+
+				b2FixtureDef fixtureDef;
+				fixtureDef.shape = &polygonShape;
+				fixtureDef.density = bc2d.density;
+				fixtureDef.friction = bc2d.friction;
+				fixtureDef.restitution = bc2d.restitution;
+				fixtureDef.restitutionThreshold = bc2d.restitutionThreshold;
+				body->CreateFixture(&fixtureDef);
+			}
+		}
+	}
+
+	// On runtime stop
+	void Scene::onRuntimeStop()
+	{
+		delete box2DWorld;
+		box2DWorld = nullptr;
+	}
+
 	// Return the registry
 	entt::registry& Scene::reg()
 	{
@@ -65,6 +126,28 @@ namespace Basil
 					
 				nsc.instance->onUpdate(timeStep);
 			});
+		}
+
+		// Physics
+		{
+			// Step the physics world
+			const int32_t velocityIterations = 6;
+			const int32_t positionIterations = 6;
+			box2DWorld->Step(timeStep, velocityIterations, positionIterations);
+
+			// Retrieve transform from Box2D, update transform component with it
+			auto view = registry.view<Rigidbody2DComponent>();
+			for (auto e : view)
+			{
+				Entity entity = { e, this };
+				auto& transform = entity.getComponent<TransformComponent>();
+				auto& rb2d = entity.getComponent<Rigidbody2DComponent>();
+				b2Body* body = (b2Body*)rb2d.runtimeBody;
+				const auto& position = body->GetPosition();
+				transform.translation.x = position.x;
+				transform.translation.y = position.y;
+				transform.rotation.z = body->GetAngle();
+			}
 		}
 
 		// Render 2D
@@ -167,5 +250,11 @@ namespace Basil
 
 	template <>
 	void Scene::onComponentAdded<TransformComponent>(Entity entity, TransformComponent& component) {}
+
+	template <>
+	void Scene::onComponentAdded<Rigidbody2DComponent>(Entity entity, Rigidbody2DComponent& component) {}
+
+	template <>
+	void Scene::onComponentAdded<BoxCollider2DComponent>(Entity entity, BoxCollider2DComponent& component) {}
 
 }